+++ packages/config/src/index.js
@@ -2,12 +2,12 @@ const { getConfigPath } = require('./path')
 const { getBuildDir } = require('./build_dir')
 const { addEnvVars } = require('./env')
 const { validateConfig } = require('./validate/main')
-const { normalizeConfig } = require('./normalize')
 const { handleFiles } = require('./files')
-const { EVENTS, LEGACY_EVENTS } = require('./events')
+const { normalizeConfig } = require('./normalize/main')
+const { EVENTS, LEGACY_EVENTS } = require('./normalize/events')
 const { parseConfig } = require('./parse/main')
 const { mergeContext } = require('./context')
-const { normalizeOpts } = require('./options')
+const { normalizeOpts } = require('./options/main')
 const { deepMerge } = require('./utils/merge')
 
 // Load the configuration file.
