+++ packages/config/src/normalize/events.js
