+++ packages/config/src/normalize/main.js
